# CaseCon: 0-bit subject check only for variable references